// src/pages/ArticlesPage/model/types/articlesPageSchema.ts
import { EntityState } from '@reduxjs/toolkit';
import { Article, ArticleView } from 'entities/Article';

export interface ArticlesPageSchema extends EntityState<Article>{
  isLoading?: boolean
  error?: string

  view: ArticleView

  // pagination
  page: number
  limit?: number
  hasMore: boolean

  _inited: boolean
}